초기 세팅 완료 2
