+++ .gitignore
@@ -32,7 +32,7 @@ venv/
 ENV/
 env/
 .venv/
-.env/
+.env
 
 # PyInstaller
 *.manifest
