+++ backend/app/config.py
@@ -1,0 +1,45 @@
+from functools import lru_cache
+from pydantic_settings import BaseSettings, SettingsConfigDict
+
+
+class Settings(BaseSettings):
+
+    # mongoDB settings
+    MONGO_URI: str
+    MONGO_ROOT_USERNAME: str
+    MONGO_ROOT_PASSWORD: str
+    # Monogo Express settings
+    MONGO_HOST: str
+    MONGO_PORT: int
+    MONGO_ADMIN_USERNAME: str
+    MONGO_ADMIN_PASSWORD: str
+    MONGO_EXPRESS_USER: str
+    MONGO_EXPRESS_PASS: str
+
+    # rabbitMQ settings
+    RABBITMQ_USER: str
+    RABBITMQ_PASS: str
+    RABBITMQ_HOST: str
+    RABBITMQ_QUEUE: str
+
+    # Redis settings
+    REDIS_HOST: str
+    REDIS_PORT: int
+
+    # Qdrant settings
+    QDRANT_HOST: str
+    QDRANT_PORT: int
+
+    # Replicate API key
+    REPLICATE_API_KEY: str
+
+    # database_password: str
+    # jwt_secret: str
+    # email_password: str
+    # celery_broker_url: str
+    # celery_backend_url: str
+
+
+@lru_cache
+def get_settings():
+    return Settings()
