+++ dev-docker-compose.yml
@@ -8,8 +8,10 @@ services:
     volumes:
       - ./backend:/backend
     environment:
-      - MONGO_URI=mongodb://admin:admin123@mongo:27017
-      - debug=True
+      - MONGO_URI=${MONGO_URI}
+      - debug=${DEBUG}
+    env_file:
+      - .env
     depends_on:
       - mongo
     networks:
@@ -31,8 +33,8 @@ services:
     ports:
       - "27017:27017"
     environment:
-      - MONGO_INITDB_ROOT_USERNAME=admin
-      - MONGO_INITDB_ROOT_PASSWORD=admin123
+      - MONGO_INITDB_ROOT_USERNAME=${MONGO_ROOT_USERNAME}
+      - MONGO_INITDB_ROOT_PASSWORD=${MONGO_ROOT_PASSWORD}
     volumes:
       - mongodb:/data/db
       - ./mongo-init:/docker-entrypoint-initdb.d
@@ -49,13 +51,13 @@ services:
     ports:
       - "8081:8081"
     environment:
-      - ME_CONFIG_MONGODB_SERVER=mongo
-      - ME_CONFIG_MONGODB_PORT=27017
-      - ME_CONFIG_MONGODB_ENABLE_ADMIN=true
-      - ME_CONFIG_MONGODB_ADMINUSERNAME=admin
-      - ME_CONFIG_MONGODB_ADMINPASSWORD=admin123
-      - ME_CONFIG_BASICAUTH_USERNAME=expressadmin
-      - ME_CONFIG_BASICAUTH_PASSWORD=express123
+      - ME_CONFIG_MONGODB_SERVER=${MONGO_HOST}
+      - ME_CONFIG_MONGODB_PORT=${MONGO_PORT}
+      - ME_CONFIG_MONGODB_ENABLE_ADMIN=${MONGO_ADMIN_ENABLE}
+      - ME_CONFIG_MONGODB_ADMINUSERNAME=${MONGO_ADMIN_USERNAME}
+      - ME_CONFIG_MONGODB_ADMINPASSWORD=${MONGO_ADMIN_PASSWORD}
+      - ME_CONFIG_BASICAUTH_USERNAME=${MONGO_EXPRESS_USER}
+      - ME_CONFIG_BASICAUTH_PASSWORD=${MONGO_EXPRESS_PASS}
     networks:
       - app-network
     hostname: mongo-express
@@ -90,8 +92,8 @@ services:
       - "5672:5672"      # 메시징 포트
       - "15672:15672"    # 웹 UI 포트
     environment:
-      - RABBITMQ_DEFAULT_USER=guest
-      - RABBITMQ_DEFAULT_PASS=guest
+      - RABBITMQ_DEFAULT_USER=${RABBITMQ_USER}
+      - RABBITMQ_DEFAULT_PASS=${RABBITMQ_PASS}
     volumes:
       - rabbitmq-data:/var/lib/rabbitmq
     networks:
