+++ pyproject.toml
@@ -9,6 +9,8 @@ dependencies = [
     "fastapi[standard]>=0.116.1",
     "passlib>=1.7.4",
     "pika>=1.3.2",
+    "pydantic>=2.11.7",
+    "pydantic-settings>=2.10.1",
     "pymongo>=4.13.2",
     "redis>=6.2.0",
 ]
